finish lora refactor replay
- Move LoraModelLoader to serve-specific module (architectural separation)
- Update all test mocking paths for new architecture
- Remove serve dependencies from common utilities (clean separation)
- Eliminate code duplication between generic & serve modules
- Fix import paths and dependency management
- Fix unit test and verify with release (probes) tests
