+++ python/ray/llm/_internal/serve/observability/usage_telemetry/usage.py
@@ -9,8 +9,8 @@ from ray._private.usage.usage_lib import (
 )
 from ray.llm._internal.common.base_pydantic import BaseModelExtended
 from ray.llm._internal.common.observability.telemetry_utils import DEFAULT_GPU_TYPE
+from ray.llm._internal.common.utils.lora_utils import get_lora_model_ids
 from ray.llm._internal.serve.observability.logging import get_logger
-from ray.llm._internal.serve.utils.lora_serve_utils import get_lora_model_ids
 
 if TYPE_CHECKING:
     from ray.llm._internal.serve.configs.server_models import LLMConfig
